Deal out of range problem

+++ Calculator.java
@@ -4,12 +4,12 @@ public class Calculator {
 	public static void main(String [] args) {
 		Scanner in = new Scanner(System.in);
 		System.out.print("Input a number: ");
-		short num1 = in.nextShort();
+		long num1 = in.nextLong();
 
 		System.out.print("Choose an operation('+','-','*','/'): ");
 		char operation = in.next().charAt(0);
 		System.out.print("Input your second number: ");
-		short num2 = in.nextShort();
+		long num2 = in.nextLong();
 		if (operation == '+') {
 			System.out.printf("Your result: %d\n",num1+num2);
 		} else if (operation == '-') {
